Move prominence swing config to detector package
