+++ ta4j-core/src/main/java/org/ta4j/core/indicators/elliott/swing/package-info.java
@@ -6,8 +6,11 @@
  *
  * <p>
  * Use this package to configure swing extraction behavior before downstream
- * Elliott scenario generation and validation. ZigZag and fractal detectors
- * cover sharp or fixed-window turns,
+ * Elliott scenario generation and validation. The types here are detector,
+ * configuration, pivot, and filter support types rather than first-class
+ * {@link org.ta4j.core.Indicator} implementations; first-class indicators keep
+ * the {@code *Indicator} naming convention. ZigZag and fractal detectors cover
+ * sharp or fixed-window turns,
  * {@link org.ta4j.core.indicators.elliott.swing.SlopeChangeSwingDetector}
  * targets rounded turns, and tolerant composite detection provides quorum
  * consensus without requiring exact pivot-index agreement.
