+++ ta4j-core/src/test/java/org/ta4j/core/indicators/elliott/swing/ProminenceSwingDetectorTest.java
@@ -7,7 +7,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 import org.ta4j.core.BarSeries;
-import org.ta4j.core.indicators.ProminenceSwingConfig;
 import org.ta4j.core.indicators.elliott.ElliottDegree;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
